requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
PyJWT==2.10.1
sqlparse==0.5.3
requests==2.31.0